Pass the graphics core to entities instead
Before this commit, PIXI app and loader were being passed to each
entity.
This commit makes it so that the GraphicsCore will just pass itself
to each Entity instead.
This is so that entities would be able to get data from other entities
as well as having the ability to spawn particles.